+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.1
+
+- Fix bug where the personality hash was not being generated correctly
+- Fix race condition with message context when character was updated
+
 # 1.1.0
 
 - Add support for sdk-core v1.1.0